simplify trial executor
